requirements.txt
discord~=1.0.1
asyncio~=3.4.3
motor~=2.3.1
requests~=2.25.0
matplotlib~=3.3.3
python-dateutil~=2.8.1